+++ src/components/diff.rs
@@ -2,7 +2,7 @@ use super::{
     CommandBlocking, Direction, DrawableComponent, ScrollType,
 };
 use crate::{
-    components::{CommandInfo, Component},
+    components::{CommandInfo, Component, EventState},
     keys::SharedKeyConfig,
     queue::{Action, InternalEvent, NeedsUpdate, Queue, ResetItem},
     strings, try_or_popup,
@@ -237,16 +237,12 @@ impl DiffComponent {
     }
 
     fn lines_count(&self) -> usize {
-        self.diff
-            .as_ref()
-            .map_or(0, |diff| diff.lines.saturating_sub(1))
+        self.diff.as_ref().map_or(0, |diff| diff.lines)
     }
 
     fn modify_selection(&mut self, direction: Direction) {
-        if let Some(diff) = &self.diff {
-            let max = diff.lines.saturating_sub(1);
-
-            self.selection.modify(direction, max);
+        if self.diff.is_some() {
+            self.selection.modify(direction, self.lines_count());
         }
     }
 
@@ -610,9 +606,11 @@ impl DrawableComponent for DiffComponent {
             r.height.saturating_sub(2),
         ));
 
+        let current_height = self.current_size.get().1;
+
         self.scroll_top.set(calc_scroll_top(
             self.scroll_top.get(),
-            self.current_size.get().1 as usize,
+            current_height as usize,
             self.selection.get_end(),
         ));
 
@@ -628,7 +626,7 @@ impl DrawableComponent for DiffComponent {
                 self.theme.text(false, false),
             )])]
         } else {
-            self.get_text(r.width, self.current_size.get().1)
+            self.get_text(r.width, current_height)
         };
 
         f.render_widget(
@@ -648,7 +646,8 @@ impl DrawableComponent for DiffComponent {
                 f,
                 r,
                 &self.theme,
-                self.lines_count(),
+                self.lines_count()
+                    .saturating_sub(usize::from(current_height)),
                 self.scroll_top.get(),
             );
         }
@@ -728,33 +727,33 @@ impl Component for DiffComponent {
     }
 
     #[allow(clippy::cognitive_complexity)]
-    fn event(&mut self, ev: Event) -> Result<bool> {
+    fn event(&mut self, ev: Event) -> Result<EventState> {
         if self.focused {
             if let Event::Key(e) = ev {
                 return if e == self.key_config.move_down {
                     self.move_selection(ScrollType::Down);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.shift_down {
                     self.modify_selection(Direction::Down);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.shift_up {
                     self.modify_selection(Direction::Up);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.end {
                     self.move_selection(ScrollType::End);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.home {
                     self.move_selection(ScrollType::Home);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.move_up {
                     self.move_selection(ScrollType::Up);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.page_up {
                     self.move_selection(ScrollType::PageUp);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.page_down {
                     self.move_selection(ScrollType::PageDown);
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.enter
                     && !self.is_immutable
                 {
@@ -764,7 +763,7 @@ impl Component for DiffComponent {
                         self.stage_unstage_hunk()
                     );
 
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.status_reset_item
                     && !self.is_immutable
                     && !self.is_stage()
@@ -776,12 +775,12 @@ impl Component for DiffComponent {
                             self.reset_hunk();
                         }
                     }
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.diff_stage_lines
                     && !self.is_immutable
                 {
                     self.stage_lines();
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.diff_reset_lines
                     && !self.is_immutable
                     && !self.is_stage()
@@ -792,17 +791,17 @@ impl Component for DiffComponent {
                             self.reset_lines();
                         }
                     }
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else if e == self.key_config.copy {
                     self.copy_selection();
-                    Ok(true)
+                    Ok(EventState::Consumed)
                 } else {
-                    Ok(false)
+                    Ok(EventState::NotConsumed)
                 };
             }
         }
 
-        Ok(false)
+        Ok(EventState::NotConsumed)
     }
 
     fn focused(&self) -> bool {