+++ src/components/mod.rs
@@ -53,6 +53,7 @@ pub use utils::filetree::FileTreeItemKind;
 use crate::ui::style::Theme;
 use anyhow::Result;
 use crossterm::event::Event;
+use std::convert::From;
 use tui::{
     backend::Backend,
     layout::{Alignment, Rect},
@@ -86,14 +87,14 @@ macro_rules! accessors {
 pub fn event_pump(
     ev: Event,
     components: &mut [&mut dyn Component],
-) -> Result<bool> {
+) -> Result<EventState> {
     for c in components {
-        if c.event(ev)? {
-            return Ok(true);
+        if c.event(ev)?.is_consumed() {
+            return Ok(EventState::Consumed);
         }
     }
 
-    Ok(false)
+    Ok(EventState::NotConsumed)
 }
 
 /// helper fn to simplify delegating command
@@ -157,6 +158,29 @@ pub trait DrawableComponent {
     ) -> Result<()>;
 }
 
+///
+#[derive(PartialEq)]
+pub enum EventState {
+    Consumed,
+    NotConsumed,
+}
+
+impl EventState {
+    pub fn is_consumed(&self) -> bool {
+        *self == Self::Consumed
+    }
+}
+
+impl From<bool> for EventState {
+    fn from(consumed: bool) -> Self {
+        if consumed {
+            Self::Consumed
+        } else {
+            Self::NotConsumed
+        }
+    }
+}
+
 /// base component trait
 pub trait Component {
     ///
@@ -166,9 +190,8 @@ pub trait Component {
         force_all: bool,
     ) -> CommandBlocking;
 
-    /// returns true if event propagation needs to end (event was consumed)
-    //TODO: lets introduce an enum `EventState` as `enum EventState { Consumed, NotConsumed }` instead of bool
-    fn event(&mut self, ev: Event) -> Result<bool>;
+    ///
+    fn event(&mut self, ev: Event) -> Result<EventState>;
 
     ///
     fn focused(&self) -> bool {