+++ src/strings.rs
@@ -71,6 +71,9 @@ pub fn commit_title_amend(_key_config: &SharedKeyConfig) -> String {
 pub fn commit_msg(_key_config: &SharedKeyConfig) -> String {
     "type commit message..".to_string()
 }
+pub fn commit_first_line_warning(count: usize) -> String {
+    format!("[subject length: {}]", count)
+}
 pub fn commit_editor_msg(_key_config: &SharedKeyConfig) -> String {
     r##"
 # Edit your commit message