+++ src/tabs/stashlist.rs
@@ -1,7 +1,7 @@
 use crate::{
     components::{
         visibility_blocking, CommandBlocking, CommandInfo,
-        CommitList, Component, DrawableComponent,
+        CommitList, Component, DrawableComponent, EventState,
     },
     keys::SharedKeyConfig,
     queue::{Action, InternalEvent, Queue},
@@ -43,7 +43,7 @@ impl StashList {
 
     ///
     pub fn update(&mut self) -> Result<()> {
-        if self.visible {
+        if self.is_visible() {
             let stashes = sync::get_stashes(CWD)?;
             let commits =
                 sync::get_commits_info(CWD, stashes.as_slice(), 100)?;
@@ -183,10 +183,13 @@ impl Component for StashList {
         visibility_blocking(self)
     }
 
-    fn event(&mut self, ev: crossterm::event::Event) -> Result<bool> {
-        if self.visible {
-            if self.list.event(ev)? {
-                return Ok(true);
+    fn event(
+        &mut self,
+        ev: crossterm::event::Event,
+    ) -> Result<EventState> {
+        if self.is_visible() {
+            if self.list.event(ev)?.is_consumed() {
+                return Ok(EventState::Consumed);
             }
 
             if let Event::Key(k) = ev {
@@ -203,7 +206,7 @@ impl Component for StashList {
             }
         }
 
-        Ok(false)
+        Ok(EventState::NotConsumed)
     }
 
     fn is_visible(&self) -> bool {